% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculoR0.R
\name{calculoR0}
\alias{calculoR0}
\title{CalculoR0}
\usage{
calculoR0(
  tabla = "//Centcs01srv03/TABLEAU/Coronavirus/R/Nuevos_positivos.csv",
  area = NULL,
  from_date = NULL,
  mean_si = 5.2,
  std_si = 4.75
)
}
\arguments{
\item{tabla}{default file location of "Nuevos_positivos.csv"}
}
\value{
R0 results total and health area specific, filtered from a specific date.Includes options to change serial interval parameters.
}
\description{
Calculates upper lower and mean values of R0
}
\examples{


}
