% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Estimate R0}
\usage{
tr(x, mean_si, std_si)
}
\arguments{
\item{x}{formated table.}
}
\value{
Estimated values of R0, upper and lower limits.
}
\description{
Estimates R0 through parametric method (EpiEstim) assuming mean serial interval=5.2 and standard dev of serial interval= 4.75
}
\examples{

}
